UHF-11306: Combine result containers, remove duplicate code
